EmailSummarizationApp_Python/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain
langchain-ollama
langchain-chroma
chromadb
pandas
bs4
